+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/NoSplineAuto/BFRightNOSplines.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Autonomous;
+package org.firstinspires.ftc.teamcode.Autonomous.NoSplineAuto;
 
 import com.acmerobotics.dashboard.config.Config;
 import com.acmerobotics.roadrunner.trajectory.constraints.AngularVelocityConstraint;
@@ -6,9 +6,9 @@ import com.acmerobotics.roadrunner.trajectory.constraints.MinVelocityConstraint;
 import com.acmerobotics.roadrunner.trajectory.constraints.TrajectoryVelocityConstraint;
 import com.acmerobotics.roadrunner.trajectory.constraints.TranslationalVelocityConstraint;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.hardware.Servo;
 
+import org.firstinspires.ftc.teamcode.Autonomous.MeepMeepBoilerplate;
 import org.firstinspires.ftc.teamcode.RoadRunner.drive.SampleMecanumDrive;
 import org.firstinspires.ftc.teamcode.Variables.Detection;
 import org.firstinspires.ftc.teamcode.Variables.VisionProcessors;
@@ -16,10 +16,8 @@ import org.firstinspires.ftc.teamcode.Variables.VisionProcessors;
 import java.util.Arrays;
 
 @Config
-@Disabled
-@Autonomous(name = "RFRightNOOOO", group = "LinearOpmode")
-public class RFRightNOSplines extends MeepMeepBoilerplate {
-
+@Autonomous(name = "BFRightNOOOO", group = "Linear OpMode")
+public class BFRightNOSplines extends MeepMeepBoilerplate {
     @Override
     public void runOpMode() {
         SampleMecanumDrive drive = new SampleMecanumDrive(hardwareMap);
@@ -39,15 +37,12 @@ public class RFRightNOSplines extends MeepMeepBoilerplate {
             telemetry.addData("Detection", detection);
             telemetry.update();
         }
-
         switch (detection) {
             case LEFT -> drive.followTrajectorySequence(
                     drive.trajectorySequenceBuilder(getCurrentPosition(drive))
-                            .back(2.0)
-                            .waitSeconds(.1)
-                            .strafeRight(11)
-                            .back(21)
-                            .waitSeconds(.1)
+                            .back(28.0)
+                            .turn(Math.toRadians(90))
+                            .back(8)
                             .addTemporalMarker(() -> passiveServo.setPosition(0.2))
                             .waitSeconds(4)
                             .forward(5)
@@ -55,7 +50,7 @@ public class RFRightNOSplines extends MeepMeepBoilerplate {
             );
             case CENTER -> { drive.followTrajectorySequence(
                     drive.trajectorySequenceBuilder(getCurrentPosition(drive))
-                            .back(32.25)
+                            .back(31.5)
                             .waitSeconds(.25)
                             .addTemporalMarker(() -> passiveServo.setPosition(0.2))
                             .waitSeconds(4)
@@ -64,9 +59,11 @@ public class RFRightNOSplines extends MeepMeepBoilerplate {
             }
             case RIGHT -> drive.followTrajectorySequence(
                     drive.trajectorySequenceBuilder(getCurrentPosition(drive))
-                            .back(28.0)
-                            .turn(Math.toRadians(-90))
-                            .back(4)
+                            .back(2.0)
+                            .waitSeconds(.25)
+                            .strafeLeft(8)
+                            .back(21)
+                            .waitSeconds(.25)
                             .addTemporalMarker(() -> passiveServo.setPosition(0.2))
                             .waitSeconds(4)
                             .forward(5)
@@ -78,6 +75,7 @@ public class RFRightNOSplines extends MeepMeepBoilerplate {
                 sleep(3000);
             }
         }
+
 
 //        drive.followTrajectorySequence(mergeSequences(sequences));
     }
