+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/NoSplineAuto/RFRightNOSplines.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Autonomous;
+package org.firstinspires.ftc.teamcode.Autonomous.NoSplineAuto;
 
 import com.acmerobotics.dashboard.config.Config;
 import com.acmerobotics.roadrunner.trajectory.constraints.AngularVelocityConstraint;
@@ -6,8 +6,10 @@ import com.acmerobotics.roadrunner.trajectory.constraints.MinVelocityConstraint;
 import com.acmerobotics.roadrunner.trajectory.constraints.TrajectoryVelocityConstraint;
 import com.acmerobotics.roadrunner.trajectory.constraints.TranslationalVelocityConstraint;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.hardware.Servo;
 
+import org.firstinspires.ftc.teamcode.Autonomous.MeepMeepBoilerplate;
 import org.firstinspires.ftc.teamcode.RoadRunner.drive.SampleMecanumDrive;
 import org.firstinspires.ftc.teamcode.Variables.Detection;
 import org.firstinspires.ftc.teamcode.Variables.VisionProcessors;
@@ -15,8 +17,10 @@ import org.firstinspires.ftc.teamcode.Variables.VisionProcessors;
 import java.util.Arrays;
 
 @Config
-@Autonomous(name = "BFRightNOSplines", group = "Linear OpMode")
-public class BFRightNOSplines extends MeepMeepBoilerplate{
+@Disabled
+@Autonomous(name = "RFRightNOOOO", group = "LinearOpmode")
+public class RFRightNOSplines extends MeepMeepBoilerplate {
+
     @Override
     public void runOpMode() {
         SampleMecanumDrive drive = new SampleMecanumDrive(hardwareMap);
@@ -36,12 +40,15 @@ public class BFRightNOSplines extends MeepMeepBoilerplate{
             telemetry.addData("Detection", detection);
             telemetry.update();
         }
+
         switch (detection) {
             case LEFT -> drive.followTrajectorySequence(
                     drive.trajectorySequenceBuilder(getCurrentPosition(drive))
-                            .back(28.0)
-                            .turn(Math.toRadians(90))
-                            .back(8)
+                            .back(2.0)
+                            .waitSeconds(.1)
+                            .strafeRight(11)
+                            .back(21)
+                            .waitSeconds(.1)
                             .addTemporalMarker(() -> passiveServo.setPosition(0.2))
                             .waitSeconds(4)
                             .forward(5)
@@ -49,7 +56,7 @@ public class BFRightNOSplines extends MeepMeepBoilerplate{
             );
             case CENTER -> { drive.followTrajectorySequence(
                     drive.trajectorySequenceBuilder(getCurrentPosition(drive))
-                            .back(31.5)
+                            .back(32.25)
                             .waitSeconds(.25)
                             .addTemporalMarker(() -> passiveServo.setPosition(0.2))
                             .waitSeconds(4)
@@ -58,11 +65,9 @@ public class BFRightNOSplines extends MeepMeepBoilerplate{
             }
             case RIGHT -> drive.followTrajectorySequence(
                     drive.trajectorySequenceBuilder(getCurrentPosition(drive))
-                            .back(2.0)
-                            .waitSeconds(.25)
-                            .strafeLeft(8)
-                            .back(21)
-                            .waitSeconds(.25)
+                            .back(28.0)
+                            .turn(Math.toRadians(-90))
+                            .back(4)
                             .addTemporalMarker(() -> passiveServo.setPosition(0.2))
                             .waitSeconds(4)
                             .forward(5)
@@ -74,7 +79,6 @@ public class BFRightNOSplines extends MeepMeepBoilerplate{
                 sleep(3000);
             }
         }
-
 
 //        drive.followTrajectorySequence(mergeSequences(sequences));
     }
