+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/SplineAuto/BCOutRightSplines.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Autonomous.Spline;
+package org.firstinspires.ftc.teamcode.Autonomous.SplineAuto;
 
 import com.acmerobotics.dashboard.config.Config;
 import com.acmerobotics.roadrunner.geometry.Pose2d;
