+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/SplineAuto/RCOutRightSplines.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Autonomous.Spline;
+package org.firstinspires.ftc.teamcode.Autonomous.SplineAuto;
 
 import com.acmerobotics.dashboard.config.Config;
 import com.acmerobotics.roadrunner.geometry.Pose2d;
@@ -19,8 +19,8 @@ import org.firstinspires.ftc.teamcode.Variables.VisionProcessors;
 import java.util.Arrays;
 
 @Config
-@Autonomous(name = "RCOutLeftSplines", group = "Linear OpMode")
-public class RCOutLeftSplines extends MeepMeepBoilerplate {
+@Autonomous(name = "RCOutRightSplines", group = "Linear OpMode")
+public class RCOutRightSplines extends MeepMeepBoilerplate {
     @Override
     public void runOpMode() {
         SampleMecanumDrive drive = new SampleMecanumDrive(hardwareMap);
@@ -70,7 +70,7 @@ public class RCOutLeftSplines extends MeepMeepBoilerplate {
                             .waitSeconds(.1)
                             .addTemporalMarker(() -> passiveServo.setPosition(0.2))
                             .waitSeconds(.5)
-                            .lineToLinearHeading(new Pose2d(51.29, -26.60, Math.toRadians(180.00)))
+                            .lineToLinearHeading(new Pose2d(51.29, -29.60, Math.toRadians(180.00)))
                             .waitSeconds(.1)
                             .addTemporalMarker(() -> autoServo.setPosition(0.35))
                             .waitSeconds(1)
@@ -86,7 +86,7 @@ public class RCOutLeftSplines extends MeepMeepBoilerplate {
                             .waitSeconds(.1)
                             .addTemporalMarker(() -> passiveServo.setPosition(0.2))
                             .waitSeconds(.5)
-                            .splineToLinearHeading(new Pose2d(51.29, -33.09, Math.toRadians(180.00)), Math.toRadians(0.00))
+                            .splineToLinearHeading(new Pose2d(51.29, -36.09, Math.toRadians(180.00)), Math.toRadians(0.00))
                             .waitSeconds(.1)
                             .addTemporalMarker(() -> autoServo.setPosition(0.35))
                             .waitSeconds(1)
@@ -103,7 +103,7 @@ public class RCOutLeftSplines extends MeepMeepBoilerplate {
                             .addTemporalMarker(() -> passiveServo.setPosition(0.2))
                             .waitSeconds(.5)
                             .lineToConstantHeading(new Vector2d(24.18, -40.99))
-                            .splineToLinearHeading(new Pose2d(51.29, -39.89, Math.toRadians(180.00)), Math.toRadians(360.00))
+                            .splineToLinearHeading(new Pose2d(51.29, -42.89, Math.toRadians(180.00)), Math.toRadians(360.00))
                             .waitSeconds(.1)
                             .addTemporalMarker(() -> autoServo.setPosition(0.35))
                             .waitSeconds(1.5)
