+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/VisionTesting/AprilTagTest1.kt
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Autonomous
+package org.firstinspires.ftc.teamcode.Autonomous.VisionTesting
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous
 import com.qualcomm.robotcore.eventloop.opmode.Disabled
